Changed the DataEntryPanelBuilder.createDataEntryPanel to accept Callable objects instead of actions.
This is because the old version of the method did odd things and would close the dialog reguardless of the other actions, also the cancel action was added to the OK button.

Now a Callable<Boolean> object needs to be passed in for both the OK and cancel buttons. If the call method returns true, the dialog will close if not it will not. The close action is no longer added to OK button (This does not seem to break any other code). Also, the CommonCloseAction is no longer used because it always closed dialog.

Some methods were changed slightly to conform with the new method paramiters. 

git-svn-id: http://sqlpower-library.googlecode.com/svn/trunk@458 49f7741a-0836-0410-9614-f5158065ea1e