Factored out the common code for refreshing the children of a SQLObject (we had almost the same code for SQLColumn and SQLIndex, and I'm about to add SQLRelationship to the mix).
This involved adding a new method to SQLObject for copying all the properties of
another object. For now, I have not made it abstract because there are some
little SQLObject implementations floating around in all the projects and I don't
want to update them. The default implementation for this non-abstract method is
to just throw an exception.

git-svn-id: http://sqlpower-library.googlecode.com/svn/trunk@801 49f7741a-0836-0410-9614-f5158065ea1e