Added catalog refresh tests to the now-inaccurately named RefreshTablesTest. In order to accomplish this, I refactored the code that fetches catalogs from the source DB to work the same as all the other SQLObject fetchXXX() methods.
Also new in this revision (created in support of the above) are the catalogTerm 
and schemaTerm properties in SQLDatabase. These are both set based on JDBC 
database metadata when the database is populated. Caching these values makes 
them much more convenient to refer to during refresh, and it's likely we'll come 
across more code that could use these instead of borrowing a connection from the 
pool just to find out these simple values.

Finally, I made a small fix to the MockJDBC stuff so that it does not return 
catalogs whose name are the empty string. This means setting "catalogs" to "" is 
the same as saying there are no catalogs at all. Formerly, this was interpreted 
to mean a single catalog whose name is the empty string.

git-svn-id: http://sqlpower-library.googlecode.com/svn/trunk@820 49f7741a-0836-0410-9614-f5158065ea1e