Added schema refresh tests to the increasingly inaccurately named RefreshTablesTest. In order to accomplish this, I refactored the code that fetches schemas from the source DB to work the same as all the other SQLObject fetchXXX() methods.
Also, I made a small fix to the MockJDBC stuff so that it does not return 
schemas whose name are the empty string. This means setting "schemas" to "" is 
the same as saying there are no schemas at all. Formerly, this was interpreted 
to mean a single schema whose name is the empty string.

git-svn-id: http://sqlpower-library.googlecode.com/svn/trunk@822 49f7741a-0836-0410-9614-f5158065ea1e