NEW - bug 2025: Add buttons for all standard functions http://trillian.sqlpower.ca/bugzilla/show_bug.cgi?id=2025
Modified the toolbars of the Query Editor, Image Viewer, and Report Editor so that they always appear at the very top, and the 'right tree' always appears under the toolbar to give more toolbar space.

I'm having some issues with doing the same with the OLAP editor, so I'm not committing that just yet.

git-svn-id: http://sqlpower-library.googlecode.com/svn/trunk@1035 49f7741a-0836-0410-9614-f5158065ea1e