Making the initial commit of SPSessionPersister.
Also changed the SPPersisterHelperFactory so that the getSPPersisterHelper now takes a simple class name instead of a Class in its parameters. This is because we do not have references to SPObject classes when we commit a persisted object. The SPPersisterHelper commitObject method signature has also been modified slightly so the PersistedSPObject comes first before the List of PersistedSPOProperty, as it seems more intuitive.

git-svn-id: http://sqlpower-library.googlecode.com/svn/trunk@1257 49f7741a-0836-0410-9614-f5158065ea1e