NEW - bug 2697: Remove the SPPersisterHelperFactoryImpl http://trillian.sqlpower.ca/bugzilla/show_bug.cgi?id=2697
The persister helpers have been split into persister helpers and secondary helpers for parent classes that are abstract. 
This is to support creating persister helpers in the Architect project. The persister helpers in Architect were coming
out empty because the annotation processor was using the abstract parent classes to aid the persister helpers in the 
library. Because the parent SPObject classes were not in Architect the helpers were empty. Now the secondary helpers are
used to replace this previous behaviour.

git-svn-id: http://sqlpower-library.googlecode.com/svn/trunk@1355 49f7741a-0836-0410-9614-f5158065ea1e