NEW - bug 2731: Undoing right after opening a server workspace is broken http://trillian.sqlpower.ca/bugzilla/show_bug.cgi?id=2731
The UndoManager can be told when it is in a loading state. This lets the UndoManager prevent events of changing the UUID 
of objects as the UUID can only change when the object is first created or during load and the UndoManager cannot undo a 
UUID while connected to a server.

The UndoManager's listener also adds a listener to ancestor nodes of objects it is listening to. This allows the UndoManager
to be aware of transaction end-points that are higher than the nodes being listened to (occurred when the root tree fired
transaction begin/end events but the UndoManager was only listening to the SQLObjectRoot).

git-svn-id: http://sqlpower-library.googlecode.com/svn/trunk@1505 49f7741a-0836-0410-9614-f5158065ea1e