NEW - bug 2748: Profling is not persisted http://trillian.sqlpower.ca/bugzilla/show_bug.cgi?id=2748
Profiling is now persisted to the server. Previously the profiles were being added to an older profile manager.
Added BigDecimal, BigInteger, Timestamp and friends to the converter for profiles to be persisted.
The Long fields in the JCR can now be used to store ints or longs. Previously, and for the most case, longs store ints and
convert the value in the JCR to an int. If the numeric field is meant to be treated as a long it will be flagged as such
since it is uncommon.

git-svn-id: http://sqlpower-library.googlecode.com/svn/trunk@1534 49f7741a-0836-0410-9614-f5158065ea1e