NEW - bug 2765: NPE when dding a column to a table that had previously been deleted http://trillian.sqlpower.ca/bugzilla/show_bug.cgi?id=2765
The root cause of this bug was the listener on the model from the table pane was being removed, then when it was revived 
from the play pen's graveyard it was no longer listening to its model. 

A side effect of this fix is the DBTreeModel's listener needed to be corrected to not pool events. This is because fixing 
the listener caused child columns to not be removable from the tables on undo. To populate SQLObjects it now works closer 
to the original implementation which added all of the objects and set populate flags, then fired events. To make updates 
from the server work a special case had to be added in to populate the SQLObjects in the same fashion that they get 
populated in the stand-alone client.

git-svn-id: http://sqlpower-library.googlecode.com/svn/trunk@1555 49f7741a-0836-0410-9614-f5158065ea1e