Modified the UserDefinedSQLType so that it uses Integer for nullability, type, and Boolean for autoIncrement, so that they can be nullable, which signifies to get the value from the upstream type. This follows the same 'inheritance' mechanism that the other properties follow, which allows for changes made in the upstream type to affect the non-overridden properties.
It is worth noting that in order to work with BeanUtils, the isAutoIncrement had to be changed to getAutoIncrement. Otherwise, BeanUtils will say that it cannot find a getter for autoIncrement, resulting in test failures.

There is a JIRA issue here about that: https://issues.apache.org/jira/browse/BEANUTILS-321

The argument for the current behaviour is that since Boolean isn't truly a boolean value (since it can be null), it should not use the isXXX format for getters.

git-svn-id: http://sqlpower-library.googlecode.com/svn/trunk@1557 49f7741a-0836-0410-9614-f5158065ea1e