Fixed some indenting on generated persister helper classes. Changed all the sb.append to the niprint, niprintln, print, and println helper methods in the SPAnnotationProcessor to keep consistent.
Also made sure that all the referenced classes in the persister helpers were generated using .class.getSimpleName() instead of using the direct string of it. This is to keep all the coding in the annotation processor consistent. As well, any renaming we might do on those classes will be much easier, and maintainable.

git-svn-id: http://sqlpower-library.googlecode.com/svn/trunk@1637 49f7741a-0836-0410-9614-f5158065ea1e