Bug fix for deadlocking on the Wabit Enterprise server. The problem was the table container was populating the tables on the server when there was a new table added to the query pen.
When using table containers on the client the table containers look up the table in the database and populate it if necessary to collect columns to use as items in the containers. Populating the table on the client also allows the client to update the table for cases where unused columns are removed or new columns are added. However, the session on the server is used to verify that the objects passed to it are valid in the system's state. We are not saving any of the tables directly as old table state is not useful for executing queries and would be a waste of time and space.

The easiest way to solve this problem is to introduce a flag to decide if the container should try to populate its model object. On the server this value is always false to prevent population and deadlocks, on clients it is always true to keep the containers up-to-date. In tests this is currently always true but could be changed.

git-svn-id: http://sqlpower-library.googlecode.com/svn/trunk@1660 49f7741a-0836-0410-9614-f5158065ea1e