Added a test to PersistedSPObjectTest to ensure that the parent property of an SPObject is only and must be set to null after firing a child removed event.
Previously, there were cases of SPObjects that set the parent property to null before firing a child removed event, which caused SPPersisterListener's propertyChanged method to throw an exception because its root object does not have a runnable dispatcher reference.

Furthermore, SPObjects that do not set the parent property to null after firing a child removed event could cause misleading and unexpected behaviour. For example, anything that still has a reference to the removed child could call getParent() and be returned a non-null value.

With this new test, all of the offending SPObjects that violate this test have been fixed.

git-svn-id: http://sqlpower-library.googlecode.com/svn/trunk@1744 49f7741a-0836-0410-9614-f5158065ea1e