Adding in control to move cross-branch relationships between same object types for proper persistence.
Years ago we realized a problem with the persister for SQLRelationship where all SQLRelationships had to be created after all SQLTables were created. However, when we create an object we normally create all descendants of an object before moving on to the next so it will be complete for future references. In this way we were making SQLRelationships before the SQLTables were all created because the relationships existed as children to the tables. To fix this at the time we put in a quick fix, specifically for SQLImportedKey in commitSortedObjects.

Now we have the same problem with Designer between Taxonomys and TaxonomyImports. Since these classes are not in the library and we want to provide a better way to handle these dependencies there is now a way to force an object order for these session persisters that makes all objects of one class be created after all objects of another class.

git-svn-id: http://sqlpower-library.googlecode.com/svn/trunk@2003 49f7741a-0836-0410-9614-f5158065ea1e